notebook-template: example template with python + r
Aim is to avoid us having too many integration issues when putting our
work together.

It loads in the data, and does the k-fold train/test splitting. It
then makes sure that data is available in both R and Python cells.

Later on, it shows how to use the predictions from a model in R to
create a confusion matrix in Python.

I believe that if we try work around this template, life will be a lot
easier at crunch time (next week).